Fix bugs
Avoid locking the application if the last subprocess produces stdout which
is not sent to a file. The output is just sent to /dev/null.
Correctly quote the working directory for tar, depending whether tar is run
in a shell or not.
Move excludes for tar to the beginning of the argument list.